Fixed player and enemy bullet stuck when die bug
This is done by handling the collision for  self.player_tank as well. I forgot that the player tank has its own variable pala.

My workaround with player bullet on respawn is that if previous player bullet still exists in the game, the new self.player_tank should acquire this bullet. Otherwise, we should just create a new bullet for the new self.player_tank.

I also switched the place for the bullet overwrite in the draw function because bullet overwrite can still happen even if on game over screen.